half term

+++ src/main/java/Controllers/Player.java
@@ -1,36 +1,60 @@
 package Controllers;
 
 import Server.Main;
+import com.sun.jersey.multipart.FormDataParam;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
+import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 
-/*@Path("Player/")*/
+@Path("Player/")
+//creates handler
 public class Player {
-    public static void insertPlayer(String Name, int WinCount) {
+    @POST
+    @Path("new")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    public String insertPlayer(@FormDataParam("Name") String Name, @FormDataParam("WinCount") Integer WinCount) {
         try {
+            if (Name == null||WinCount == null){
+                throw new Exception("One or more parameters missing in HTTP request.");
+            }
+            System.out.println("player/new Name =" + Name);
             PreparedStatement ps = Main.db.prepareStatement("INSERT INTO Players (Players.Name, Players.WinCount) VALUES (?,?)");
             ps.setString(1, Name);
             ps.setInt(2,WinCount);
             ps.executeUpdate();
             System.out.println();
+            return "{\"status\":\"OK\"}";
         } catch (Exception exception) {
             System.out.println("Error: "+ exception.getMessage());
+            return "{\"error\":\"Unable to add player, see server info for more information\"}";
         }
     }
-    public static void deletePlayer(String Name){
+    @POST
+    @Path("delete")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    //creates handler for /players/delete
+    public String deletePlayer(@FormDataParam("Name") String Name){
         try{
+            if (Name == null){
+                throw new Exception("One or more parameters missing in HTTP request.");
+            }
+            //specifies parameters to pass into API method
+            System.out.println("player/delete name=" + Name);
             PreparedStatement ps = Main.db.prepareStatement("DELETE FROM Players WHERE Players.Name = ?");
             ps.setString(1, Name);
             ps.executeUpdate();
+            //deletes player from database
+            return "{\"status\": \"OK\"}";
         }catch (Exception exception) {
             System.out.println("Error: "+ exception.getMessage());
+            return "{\"error\":\"Unable to delete, see server console for more information\"}";
+            //reports error if unable to complete update
         }
     }
     @GET
@@ -55,14 +79,29 @@ public class Player {
             return "{\"error\": \"Unable to list items, please see server console for more info.\"}";
         }
     }
-    public static void updatePlayer(String playerName, int playerID){
+    @POST
+    @Path("update")
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
+    @Produces(MediaType.APPLICATION_JSON)
+    //creates handler for /players/update
+    public String updatePlayer(@FormDataParam("playerName") String playerName, @FormDataParam("playerID") Integer playerID){
+       //specifies parameters to pass into API method
         try {
+            if(playerName == null || playerID == null) {
+                throw new Exception("One or more parameters missing in HTTP request.");
+            }
+            //ensures the request will not  fail due to empty parameters
+            System.out.println("player/update name=" + playerName);
             PreparedStatement ps = Main.db.prepareStatement("UPDATE Players SET Players.Name = ? WHERE Players.PlayerID = ?");
             ps.setString(1, playerName);
             ps.setInt(2, playerID);
             ps.executeUpdate();
+            //updates player details
+            return "{\"status\": \"OK\"}";
         } catch (Exception exception) {
             System.out.println("Error: " + exception.getMessage());
+            return "{\"error\":\"Unable to update, please see server console for more information}";
+            //reports error if unable to complete update
         }
 
     }
